refactor: remove remaining dead code suppressions
