+++ fuzz/fuzz_targets/formatter_validity_fuzz.rs
@@ -3,6 +3,7 @@
 #![no_main]
 
 mod common;
+mod formatter_validity_common;
 
 use libfuzzer_sys::{Corpus, fuzz_target};
 use shuck_formatter::{FormatError, format_source};
@@ -13,22 +14,27 @@ fuzz_target!(|data: &[u8]| -> Corpus {
         Err(reject) => return reject,
     };
 
-    for case in common::FORMAT_CASES {
+    for case in formatter_validity_common::FORMAT_CASES {
         let path = case.path();
         let options = case.format_options();
 
         let formatted = match format_source(input, Some(path), &options) {
-            Ok(result) => common::format_result_to_string(result, input),
+            Ok(result) => formatter_validity_common::format_result_to_string(result, input),
             Err(FormatError::Parse { .. }) => continue,
             Err(FormatError::Internal(message)) => {
                 panic!("internal formatter error for {}: {message}", path.display())
             }
         };
 
-        let original_diagnostics = common::lint_source_strict(input, path, case.parse_dialect());
+        let original_diagnostics =
+            formatter_validity_common::lint_source_strict(input, path, case.parse_dialect());
         let formatted_diagnostics =
-            common::lint_source_strict(&formatted, path, case.parse_dialect());
-        common::compare_lint_counts(&original_diagnostics, &formatted_diagnostics, path);
+            formatter_validity_common::lint_source_strict(&formatted, path, case.parse_dialect());
+        formatter_validity_common::compare_lint_counts(
+            &original_diagnostics,
+            &formatted_diagnostics,
+            path,
+        );
     }
 
     Corpus::Keep
