+++ fuzz/fuzz_targets/linter_no_panic_fuzz.rs
@@ -3,6 +3,7 @@
 #![no_main]
 
 mod common;
+mod linter_common;
 
 use libfuzzer_sys::{Corpus, fuzz_target};
 
@@ -12,11 +13,13 @@ fuzz_target!(|data: &[u8]| -> Corpus {
         Err(reject) => return reject,
     };
 
-    for case in common::FORMAT_CASES {
+    for case in linter_common::LINT_CASES {
         let path = case.path();
 
-        let with_path = common::lint_source_with_recovery(input, Some(path), case.parse_dialect());
-        let without_path = common::lint_source_with_recovery(input, None, case.parse_dialect());
+        let with_path =
+            linter_common::lint_source_with_recovery(input, Some(path), case.parse_dialect());
+        let without_path =
+            linter_common::lint_source_with_recovery(input, None, case.parse_dialect());
 
         for diagnostic in with_path.iter().chain(without_path.iter()) {
             assert!(
@@ -24,7 +27,7 @@ fuzz_target!(|data: &[u8]| -> Corpus {
                 "linter emitted an empty diagnostic message for {}",
                 path.display()
             );
-            common::assert_span_valid(diagnostic.span, input);
+            linter_common::assert_span_valid(diagnostic.span, input);
         }
     }
